fixed submit code
