+++ petbook_fe/components/login/styled/styledLoginSubmit.tsx
@@ -2,6 +2,35 @@ import styled from "styled-components";
 
 const ButtonBox = styled.div`
   margin-top: 42px;
+  a {
+    display: block;
+    width: 100%;
+    padding: 20px 0;
+    margin-bottom: 12px;
+    border-radius: 12px;
+    text-align: center;
+    font-weight: 700;
+    font-size: 20px;
+    color: white;
+    box-sizing: border-box;
+    &:last-child {
+      margin-bottom: 0;
+    }
+    &.email {
+      padding: 28px 0;
+      margin-bottom: 60px;
+      background-color: var(--main);
+    }
+    &.naver {
+      background-color: #41d97e;
+    }
+    &.kakao {
+      background-color: #ffc700;
+    }
+    &.google {
+      background-color: #7270ff;
+    }
+  }
 `;
 const Container = styled.div`
   padding: 41px 35px;
@@ -23,4 +52,41 @@ const Container = styled.div`
   }
 `;
 
-export { ButtonBox, Container };
+const PassGuide = styled.figure`
+  margin-top: 28px;
+  text-align: center;
+  p {
+    display: inline;
+    font-size: 18px;
+  }
+  a {
+    display: inline;
+    padding: 0;
+    margin-left: 8px;
+    font-size: 18px;
+    font-weight: 400;
+    color: #111;
+    text-decoration: underline;
+  }
+`;
+
+const LoginWrap = styled.div`
+  position: relative;
+  width: 679px;
+  margin: 0 auto;
+  margin-top: 194px;
+`;
+
+const NotLogin = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 55px;
+  font-size: 20px;
+  color: #b6b6b6;
+  svg {
+    margin-bottom: 25px;
+    font-size: 55px;
+  }
+`;
+export { ButtonBox, Container, PassGuide, LoginWrap, NotLogin };
